# api/doc/swagger.yml
openapi: "3.0.1"
info: 
  title: "API GPAO"
  version: "1.0.0"
  description: "Documentation de l'API mise en place dans le cadre de la refonte de la GPAO."
servers:
  - url: "http://localhost:8080/api"
    description: "Serveur de production"
...
paths:
...
  '/job/{id}':
    get:
      tags:
        - jobs
      summary: "Récupération d'un job en particulier"
      description: "Récupération de toutes les informations relatives à un job en fonction de son id"
      parameters:
        - in: path
          name: id
          description: l'identifiant du job
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...